DEV: Clean up hashtag code
* Delete dead code
* Split up hashtag-autocomplete into more logical modules
