+++ app/assets/javascripts/discourse/lib/hashtag-type-registry.js
@@ -1,0 +1,10 @@
+let hashtagTypeClasses = {};
+export function registerHashtagType(type, typeClassInstance) {
+  hashtagTypeClasses[type] = typeClassInstance;
+}
+export function cleanUpHashtagTypeClasses() {
+  hashtagTypeClasses = {};
+}
+export function getHashtagTypeClasses() {
+  return hashtagTypeClasses;
+}
