+++ plugins/chat/assets/javascripts/discourse/lib/transform-auto-links.js
@@ -1,4 +1,4 @@
-import { generatePlaceholderHashtagHTML } from "discourse/lib/hashtag-autocomplete";
+import { generatePlaceholderHashtagHTML } from "discourse/lib/hashtag-decorator";
 import getURL from "discourse-common/lib/get-url";
 
 const domParser = new DOMParser();
